Added "run-calibration-test" script.
This gives (rather anecdotal) timing information for the planner on a
variety of benchmarks and configurations.

Intended for checking out if our performance on Mac OS is OK.

This is simply run on the command line, without using the experiment
infrastructure or requiring any special computation structure. Including
compilation, this should run for less than an hour and will only occupy
a single core.

As always, it is recommended to keep the machine quiet while benchmarking.