Fix bugs in normalization.
Variables in conditions are now always uniquified and simplified
upon parsing. Uniquification used to happen later on, but not for
all kinds of conditions (for example not for the goal), and
simplification would also not happen for all kinds of conditions
(for example not for the goal).

It is possible that the new code now uniquifies certain conditions
multiple times, but there is no harm in that other than a slight
performance cost.